+++ msgpack_numpy.py
@@ -87,7 +87,7 @@ def decode(obj):
                 return np.fromstring(obj[b'data'],
                             dtype=np.dtype(descr))[0]
         elif b'complex' in obj:
-            return complex(obj[b'data'])
+            return complex(tostr(obj[b'data']))
         else:
             return obj
     except KeyError:
@@ -105,7 +105,7 @@ if msgpack.version < (0, 4, 0):
                                          encoding=encoding,
                                          unicode_errors=unicode_errors,
                                          use_single_float=use_single_float,
-                                         autoreset=1)
+                                         autoreset=autoreset)
     class Unpacker(_unpacker.Unpacker):
         def __init__(self, file_like=None, read_size=0, use_list=None,
                      object_hook=decode,
@@ -128,18 +128,18 @@ else:
                      unicode_errors='strict',
                      use_single_float=False,
                      autoreset=1,
-                     use_bin_type=1):
+                     use_bin_type=0):
             super(Packer, self).__init__(default=default,
                                          encoding=encoding,
                                          unicode_errors=unicode_errors,
                                          use_single_float=use_single_float,
-                                         autoreset=1,
-                                         use_bin_type=1)
+                                         autoreset=autoreset,
+                                         use_bin_type=use_bin_type)
 
     class Unpacker(_unpacker.Unpacker):
         def __init__(self, file_like=None, read_size=0, use_list=None,
                      object_hook=decode,
-                     object_pairs_hook=None, list_hook=None, encoding='utf-8',
+                     object_pairs_hook=None, list_hook=None, encoding=None,
                      unicode_errors='strict', max_buffer_size=0,
                      ext_hook=msgpack.ExtType):
             super(Unpacker, self).__init__(file_like=file_like,
@@ -170,21 +170,21 @@ def packb(o, default=encode, **kwargs):
     kwargs['default'] = default
     return Packer(**kwargs).pack(o)
 
-def unpack(stream, object_hook=decode, encoding='utf-8', **kwargs):
+def unpack(stream, object_hook=decode, **kwargs):
     """
     Unpack a packed object from a stream.
     """
 
     kwargs['object_hook'] = object_hook
-    return _unpacker.unpack(stream, encoding=encoding, **kwargs)
+    return _unpacker.unpack(stream, **kwargs)
 
-def unpackb(packed, object_hook=decode, encoding='utf-8', **kwargs):
+def unpackb(packed, object_hook=decode, **kwargs):
     """
     Unpack a packed object.
     """
 
     kwargs['object_hook'] = object_hook
-    return _unpacker.unpackb(packed, encoding=encoding, **kwargs)
+    return _unpacker.unpackb(packed, **kwargs)
 
 load = unpack
 loads = unpackb
@@ -214,109 +214,189 @@ if __name__ == '__main__':
         pass # Python 3
 
     from unittest import main, TestCase, TestSuite
-
+    from numpy.testing import assert_equal, assert_array_equal
+    
     class test_numpy_msgpack(TestCase):
         def setUp(self):
              patch()
-        def encode_decode(self, x):
-            x_enc = msgpack.packb(x)
-            return msgpack.unpackb(x_enc)
+             
+        def encode_decode(self, x, use_bin_type=False, encoding=None):
+            x_enc = msgpack.packb(x, use_bin_type=use_bin_type)
+            return msgpack.unpackb(x_enc, encoding=encoding)
+        
+        def test_bin(self):
+            # Since bytes == str in Python 2.7, the following
+            # should pass on both 2.7 and 3.*
+            assert_equal(type(self.encode_decode(b'foo')), bytes)
+                
+        def test_str(self):
+            assert_equal(type(self.encode_decode('foo')), bytes)
+            if sys.version_info.major == 2:
+                assert_equal(type(self.encode_decode(u'foo')), str)
+
+                # Test non-default string encoding/decoding:
+                assert_equal(type(self.encode_decode(u'foo', True, 'utf=8')), unicode)
+                
         def test_numpy_scalar_bool(self):
             x = np.bool_(True)
             x_rec = self.encode_decode(x)
-            assert x == x_rec and type(x) == type(x_rec)
+            assert_equal(x, x_rec)
+            assert_equal(type(x), type(x_rec))
             x = np.bool_(False)
             x_rec = self.encode_decode(x)
-            assert x == x_rec and type(x) == type(x_rec)
+            assert_equal(x, x_rec)
+            assert_equal(type(x), type(x_rec))
+            
         def test_numpy_scalar_float(self):
             x = np.float32(np.random.rand())
             x_rec = self.encode_decode(x)
-            assert x == x_rec and type(x) == type(x_rec)
+            assert_equal(x, x_rec)
+            assert_equal(type(x), type(x_rec))
+            
         def test_numpy_scalar_complex(self):
             x = np.complex64(np.random.rand()+1j*np.random.rand())
             x_rec = self.encode_decode(x)
-            assert x == x_rec and type(x) == type(x_rec)
+            assert_equal(x, x_rec)
+            assert_equal(type(x), type(x_rec))
+            
         def test_scalar_float(self):
             x = np.random.rand()
             x_rec = self.encode_decode(x)
-            assert x == x_rec and type(x) == type(x_rec)
+            assert_equal(x, x_rec)
+            assert_equal(type(x), type(x_rec))
+            
         def test_scalar_complex(self):
             x = np.random.rand()+1j*np.random.rand()
             x_rec = self.encode_decode(x)
-            assert x == x_rec and type(x) == type(x_rec)
+            assert_equal(x, x_rec)
+            assert_equal(type(x), type(x_rec))
+            
         def test_list_numpy_float(self):
             x = [np.float32(np.random.rand()) for i in range(5)]
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x, y: x == y, x, x_rec)) and all(map(lambda x,y: type(x) == type(y), x, x_rec))
+            assert_array_equal(x, x_rec)
+            assert_array_equal([type(e) for e in x],
+                               [type(e) for e in x_rec])
+            
         def test_list_numpy_float_complex(self):
             x = [np.float32(np.random.rand()) for i in range(5)] + \
               [np.complex128(np.random.rand()+1j*np.random.rand()) for i in range(5)]
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x, x_rec)) and all(map(lambda x,y: type(x) == type(y), x, x_rec))
+            assert_array_equal(x, x_rec)
+            assert_array_equal([type(e) for e in x],
+                               [type(e) for e in x_rec])
+            
         def test_list_float(self):
             x = [np.random.rand() for i in range(5)]
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x, x_rec)) and all(map(lambda x,y: type(x) == type(y), x, x_rec))
+            assert_array_equal(x, x_rec)
+            assert_array_equal([type(e) for e in x],
+                               [type(e) for e in x_rec])
+            
         def test_list_float_complex(self):
             x = [(np.random.rand()+1j*np.random.rand()) for i in range(5)]
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x, y: x == y, x, x_rec)) and all(map(lambda x,y: type(x) == type(y), x, x_rec))
+            assert_array_equal(x, x_rec)
+            assert_array_equal([type(e) for e in x],
+                               [type(e) for e in x_rec])
+            
         def test_list_str(self):
-            x = ['x'*i for i in range(5)]
+            x = [b'x'*i for i in range(5)]
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x, x_rec)) and all(map(lambda x,y: type(x) == type(y), x, x_rec))
+            assert_array_equal(x, x_rec)
+            assert_array_equal([type(e) for e in x_rec], [bytes]*5)
+            
         def test_dict_float(self):
-            x = {'foo': 1.0, 'bar': 2.0}
+            x = {b'foo': 1.0, b'bar': 2.0}
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x.values(), x_rec.values())) and \
-                           all(map(lambda x,y: type(x) == type(y), x.values(), x_rec.values()))
+            assert_array_equal(sorted(x.values()), sorted(x_rec.values()))
+            assert_array_equal([type(e) for e in sorted(x.values())],
+                               [type(e) for e in sorted(x_rec.values())])
+            assert_array_equal(sorted(x.keys()), sorted(x_rec.keys()))
+            assert_array_equal([type(e) for e in sorted(x.keys())],
+                               [type(e) for e in sorted(x_rec.keys())])
+            
         def test_dict_complex(self):
-            x = {'foo': 1.0+1.0j, 'bar': 2.0+2.0j}
+            x = {b'foo': 1.0+1.0j, b'bar': 2.0+2.0j}
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x.values(), x_rec.values())) and \
-                           all(map(lambda x,y: type(x) == type(y), x.values(), x_rec.values()))
+            assert_array_equal(sorted(x.values(), key=np.linalg.norm),
+                               sorted(x_rec.values(), key=np.linalg.norm))
+            assert_array_equal([type(e) for e in sorted(x.values(), key=np.linalg.norm)],
+                               [type(e) for e in sorted(x_rec.values(), key=np.linalg.norm)])
+            assert_array_equal(sorted(x.keys()), sorted(x_rec.keys()))
+            assert_array_equal([type(e) for e in sorted(x.keys())],
+                               [type(e) for e in sorted(x_rec.keys())])
+
         def test_dict_str(self):
-            x = {'foo': 'xxx', 'bar': 'yyyy'}
+            x = {b'foo': b'xxx', b'bar': b'yyyy'}
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x.values(), x_rec.values())) and \
-                           all(map(lambda x,y: type(x) == type(y), x.values(), x_rec.values()))
+            assert_array_equal(sorted(x.values()), sorted(x_rec.values()))
+            assert_array_equal([type(e) for e in sorted(x.values())],
+                               [type(e) for e in sorted(x_rec.values())])
+            assert_array_equal(sorted(x.keys()), sorted(x_rec.keys()))
+            assert_array_equal([type(e) for e in sorted(x.keys())],
+                               [type(e) for e in sorted(x_rec.keys())])
+            
         def test_dict_numpy_float(self):
-            x = {'foo': np.float32(1.0), 'bar': np.float32(2.0)}
+            x = {b'foo': np.float32(1.0), b'bar': np.float32(2.0)}
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x.values(), x_rec.values())) and \
-                           all(map(lambda x,y: type(x) == type(y), x.values(), x_rec.values()))
+            assert_array_equal(sorted(x.values()), sorted(x_rec.values()))
+            assert_array_equal([type(e) for e in sorted(x.values())],
+                               [type(e) for e in sorted(x_rec.values())])
+            assert_array_equal(sorted(x.keys()), sorted(x_rec.keys()))
+            assert_array_equal([type(e) for e in sorted(x.keys())],
+                               [type(e) for e in sorted(x_rec.keys())])
+
         def test_dict_numpy_complex(self):
-            x = {'foo': np.complex128(1.0+1.0j), 'bar': np.complex128(2.0+2.0j)}
+            x = {b'foo': np.complex128(1.0+1.0j), b'bar': np.complex128(2.0+2.0j)}
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x.values(), x_rec.values())) and \
-                           all(map(lambda x,y: type(x) == type(y), x.values(), x_rec.values()))
+            assert_array_equal(sorted(x.values(), key=np.linalg.norm),
+                               sorted(x_rec.values(), key=np.linalg.norm))
+            assert_array_equal([type(e) for e in sorted(x.values(), key=np.linalg.norm)],
+                               [type(e) for e in sorted(x_rec.values(), key=np.linalg.norm)])
+            assert_array_equal(sorted(x.keys()), sorted(x_rec.keys()))
+            assert_array_equal([type(e) for e in sorted(x.keys())],
+                               [type(e) for e in sorted(x_rec.keys())])
+            
         def test_numpy_array_float(self):
             x = np.random.rand(5).astype(np.float32)
             x_rec = self.encode_decode(x)
-            assert np.all(x == x_rec) and x.dtype == x_rec.dtype
+            assert_array_equal(x, x_rec)
+            assert_equal(x.dtype, x_rec.dtype)
+            
         def test_numpy_array_complex(self):
             x = (np.random.rand(5)+1j*np.random.rand(5)).astype(np.complex128)
             x_rec = self.encode_decode(x)
-            assert np.all(x == x_rec) and x.dtype == x_rec.dtype
+            assert_array_equal(x, x_rec)
+            assert_equal(x.dtype, x_rec.dtype)
+            
         def test_numpy_array_float_2d(self):
             x = np.random.rand(5,5).astype(np.float32)
             x_rec = self.encode_decode(x)
-            assert np.all(x == x_rec) and x.dtype == x_rec.dtype
+            assert_array_equal(x, x_rec)
+            assert_equal(x.dtype, x_rec.dtype)
+            
         def test_numpy_array_str(self):
-            x = np.array(['aaa', 'bbbb', 'ccccc'])
+            x = np.array([b'aaa', b'bbbb', b'ccccc'])
             x_rec = self.encode_decode(x)
-            assert np.all(x == x_rec) and x.dtype == x_rec.dtype
+            assert_array_equal(x, x_rec)
+            assert_equal(x.dtype, x_rec.dtype)
+            
         def test_numpy_array_mixed(self):
-            x = np.array([(1, 2, 'a')],
+            x = np.array([(1, 2, b'a')],
                          np.dtype([('arg0', np.uint32),
                                    ('arg1', np.uint32),
                                    ('arg2', 'S1')]))
             x_rec = self.encode_decode(x)
-            assert np.all(x == x_rec) and x.dtype == x_rec.dtype
+            assert_array_equal(x, x_rec)
+            assert_equal(x.dtype, x_rec.dtype)
+            
         def test_list_mixed(self):
-            x = [1.0, np.float32(3.5), np.complex128(4.25), 'foo']
+            x = [1.0, np.float32(3.5), np.complex128(4.25), b'foo']
             x_rec = self.encode_decode(x)
-            assert all(map(lambda x,y: x == y, x, x_rec)) and all(map(lambda x,y: type(x) == type(y), x, x_rec))
+            assert_array_equal(x, x_rec)
+            assert_array_equal([type(e) for e in x],
+                               [type(e) for e in x_rec])
 
     main()